Implementing the new annotation writing class, and updating the standard .ann results to include the new statistics information. All looking good.
git-svn-id: http://svn.atnf.csiro.au/duchamp/trunk@1074 450e7745-72da-457f-8a74-b07dc0cda934

+++ src/Outputs/AnnotationWriter.cc
@@ -127,7 +127,7 @@ namespace duchamp {
       if(this->itsParam->getAnnotationType() == "borders"){
 	double *pix = new double[3];
 	double *wld = new double[3];
-	float x1,x2,y1,y2;
+	double x1,x2,y1,y2;
 	std::vector<int> vertexSet = object->getVertexSet();
 	for(size_t i=0;i<vertexSet.size()/4;i++){
 	  pix[0] = vertexSet[i*4]-0.5;
